update docs to reflect runners-are-stateless-and-are-the-plugins
biggest change is that multiple sessions are no longer really needed
as one can just create another RunnerManager and access everything
through that. essentially the RunnreManager *is* the session. i spent
some time this morning mapping what would need to go into the session
tracking object and it came down to "everything, except the runner
instances". so runners are now required to be stateless and RunnerManager
can be the session.

i'm actually tempted to rename the class at this point