fix updateMatches, use size() (rather than count()) consistently
updateMatches now records which matches are updated and calls for a
data sync.

in syncMatches, the model is informed of ALL updated matches first
before syncronizing new matches can happen.

in theory this may waste some cycles since an updated match may get
immediately overwritten; in practice this should not matter much as
syncMatches is run in the main thread and so no painting should be
caused as a result.

in practice, this gives us 100% sure-to-happen updates: they are
always applied in the correct thread before the set of syncedMatches
is changed