unbreak QuerySession::halt()
it was clearing all the RunnerSessionData objects (good) but they
were ONLY getting set up on runner load (bad).

so now we track if the session data has actually been retrieved or
not in that session in the RunnerMetaData object. this allows
retrieval post-plugin-load but without repeatedly calling it in the
case where a runner fails to return a RunnerSessionData object