turn QuerySessionThread into a QObject and put it into a QThread with moveToThread
this simplifies the signal/slot maze somewhat. introduced a "which thread are we in"
debug #define to assert we're in the correct thread at all times.

the header for the QuerySessionThread was also divided up by where the methods are
expecting to be called from. as this is all internal API, this is acceptable.