make QueryMatch hold the SessionData, and have RunnerSessionData set it
previously QueryMatch was keeping track of its associated runner. but
matches belong sessions, really, and from a sessiondata object the
runner can be retrieved. this also allows runners to use the
sessiondata in exec()

where this may cause issues in future is in saving/restoring QueryMatch
objects and then exec'ing them. something we did iirc in Plasma Netbook's
Search and Launch. something to replace that functionality may well be
needed in future, and should probably be placed in QuerySession itself
to avoid ad-hoc hacks.